code is running, need to get saving portion running

+++ Projects/Collaborate_Compete/buffers.py
@@ -113,7 +113,7 @@ class ReplayBuffer:
             # store the current state, current action, cumulative discounted
             # reward from t -> t+n-1, and the next_state at t+n (S't+n)
             observations = observations[0]
-            actions = torch.from_numpy(actions[0])
+            actions = torch.from_numpy(actions[0]).type(torch.float)
             summed_rewards = torch.tensor([summed_rewards]).type(torch.float)
             next_observations = next_observations[n_steps]
             dones = torch.tensor([dones[n_steps]])
